+++ src/core/javascripts/services/modal_form.js
@@ -119,7 +119,7 @@
                     $scope.form_model.datetime = moment($scope.form_model.datetime).tz(CompanyStoreService.time_zone);
                 }*///TODO double check
 
-                $scope.form_model.datetime = bbTimeZone.convertToCompanyTz($scope.form_model.datetime);
+                $scope.form_model.datetime = bbTimeZone.convertToCompany($scope.form_model.datetime);
 
 
                 $scope.model.$update($scope.form_model).then(function () {