+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to this project will be documented in this file using [CHANG
 
 ### Changed
 * GeneralOptions.use_local_time_zone changed to bbi18nOptions.use_browser_time_zone
-* GeneralOptions.display_time_zone changed to bbTimeZone.getDisplayTimeZone()
+* GeneralOptions.display_time_zone changed to bbTimeZone.getDisplay()
   
 * BREAKING: bbAdminBookingClients directive is not exposing ValidatorService anymore to the view: $scope.validator. For any bespoke project which overrides template - `admin_booking_clients.html`, replace `validator.getEmailPattern()` with `emailPattern`.